+++ junit5-jupiter-extensions/src/test/java/com/example/singleton/Builder123.java
@@ -1,30 +1,9 @@
 package com.example.singleton;
 
-import com.example.singleton.SingletonParameterResolver.Resource;
+public class Builder123 extends AbstractStringBuilderResource {
 
-public class Builder123 implements Resource<StringBuilder> {
-
-	private final StringBuilder instance;
-
-	public Builder123(/*org.junit.jupiter.api.extension.ExtensionContext context*/) {
-		this.instance = new StringBuilder("123");
-		log(">> created");
+	public Builder123() {
+		super(new StringBuilder("123"));
 	}
 
-	@Override
-	public void close() {
-		// Resource.super.close();
-		log("<< closed");
-	}
-
-	@Override
-	public StringBuilder getInstance() {
-		return instance;
-	}
-
-	private void log(String message) {
-		StringBuilder builder = getInstance();
-		String identity = "0x" + Integer.toHexString(System.identityHashCode(builder)).toUpperCase();
-		System.out.println(identity + ": " + message);
-	}
 }
