+++ junit5-jupiter-extensions/src/test/java/com/example/singleton/BuilderWithContextAccess.java
@@ -1,0 +1,11 @@
+package com.example.singleton;
+
+import org.junit.jupiter.api.extension.ExtensionContext;
+
+public class BuilderWithContextAccess extends AbstractStringBuilderResource {
+
+	public BuilderWithContextAccess(ExtensionContext context) {
+		super(new StringBuilder(context.getDisplayName()));
+	}
+
+}
