+++ junit5-jupiter-extensions/src/test/java/com/example/singleton/SingletonParameterDemoTests.java
@@ -23,7 +23,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
  * @see SingletonParameterResolver
  */
 @ExtendWith(SingletonParameterResolver.class)
-class SingletonParameterResolverTests {
+class SingletonParameterDemoTests {
 
 	private void log(String message, StringBuilder builder, TestInfo info) {
 		String identity = "0x" + Integer.toHexString(System.identityHashCode(builder)).toUpperCase();
